Rename project1.py to Job Recommendation System.py
A recommendation system is a popular application of Data Science that recommends personalized content based on the users’ interests. Almost all the popular websites you visit today use a recommendation system. As the name suggests, a job recommendation system is an application that recommends jobs based on the skills and the user’s desired role.